refactor: Summary 관련 폴더 구조 정리
